multi request alpha.1

+++ src/nodeserver/express.sh
@@ -8,11 +8,13 @@ startSecond="$3"
 endDate="$4"
 endTime="$5"
 endSecond="$6"
+restant="$7"
 
 #creation du fichier de la journee correspondante
 echo "c'est parti" 
 echo "$startDate"
-filelist=$(find  -name "trace-$1*.zip" -exec sh -c 'unzip -jq -d tmp {}' ';')
+
+filelist=$(find  -name "trace-$1*.zip" -exec sh -c 'unzip -jqn -d tmp {}' ';')
 for filename in ./tmp/*.pcap; do
 	pcapfix -dn "$filename"
 done
@@ -28,8 +30,10 @@ then
 	editcap -A "$year-$1 $2:$3"  -B "$year-$4 $5:$6" input"$year-$1 $2:$3 $year-$4 $5:$6".pcap /root/ftpuser/"$year-$1 $2:$3 $year-$4 $5:$6".pcap  &&
 	echo "edit termine" 
 	#nettoyage
-	rm *.pcap
-	rm -Rf ./tmp &&
+    if [ "$restant" == "0"]; then
+	    rm *.pcap
+	    rm -Rf ./tmp &&
+    fi
 	echo "operation termine"
 else
 	echo "pas de fichier trouve"
